Added graph class and graph_visualize method
The graph class provides a representation of both directed and
undirected graphs. Example
g = graph(4, set(()), False)
Is the undirected graph with 4 vertices and no edges.
g = graph(5, ((1, 2), (2, 3)), True)
Is the directed graph on 5 vertices with edges 1 -> 2, 2 <- 1.
graph_visualize(g) prints a representation of the graph in a form
that can be compiled using tikz.